[graph]: Pruning computations fixed logic not right as I/O not being performed for some that should

+++ flash-graph/libgraph-algs/sem_kmeans.h
@@ -169,7 +169,7 @@ namespace {
                 complete = false;
 
                 for (unsigned i = 0; i < size(); i++) {
-                    this->mean[i] *= num_members;
+                    this->mean[i] *= (double)num_members;
                 }
             }
